refactor(764): Sortir des composants pour les infos offres et info PJ
